Move media block type enum to media upload completion processor
